feat: Separate JoinGame method from main function
