+++ cmd/server/main.go
@@ -1,37 +1,17 @@
 package main
 
 import (
-	"context"
 	"fmt"
 	"net/http"
 
-	"connectrpc.com/connect"
-	apiv1 "github.com/JanStanleyWatt/testing-cardgame/dist/autogen/go/api/v1"
+	"github.com/JanStanleyWatt/testing-cardgame/cmd/server/logic"
 	"github.com/JanStanleyWatt/testing-cardgame/dist/autogen/go/api/v1/apiv1connect"
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
-// Empty struct for GameServer
-// この空構造体にGameServiceHandlerインターフェースを実装する
-type GameServer struct{}
-
-// JoinGame is a method of GameServer
-func (s *GameServer) JoinGame(
-	ctx context.Context, req *connect.Request[apiv1.JoinGameRequest]) (*connect.Response[apiv1.JoinGameResponse], error) {
-
-	res := connect.NewResponse(&apiv1.JoinGameResponse{
-		Success:      true,
-		Message:      "Welcome!",
-		ResponseTime: timestamppb.Now(),
-	})
-
-	return res, nil
-}
-
 func main() {
-	gameserver := &GameServer{}
+	gameserver := &logic.GameServer{}
 	mux := http.NewServeMux()
 
 	path, handler := apiv1connect.NewGameServiceHandler(gameserver)
